backend/requirements.txt
fastapi
uvicorn
soundcloud-lib
aiohttp
python-multipart
pywebview
requests
pyinstaller

